Video_Player/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
opencv-python==4.8.1.78
PyQt5==5.15.10
keras==2.15.0
tensorflow==2.15.0
pillow==10.1.0
requests==2.31.0
numpy==1.24.3
python-multipart==0.0.6
